Product 도메인 모델 수정
